+++ tests/test_biophysical/test_crop_soil_field/field_tests/test_manure_application.py
@@ -3,10 +3,12 @@ from unittest.mock import MagicMock, call, patch
 
 import pytest
 from pytest import approx
+from pytest_mock import MockerFixture
 
 from RUFAS.biophysical.field.field.manure_application import ManureApplication
 from RUFAS.biophysical.field.soil.manure_pool import ManurePool
 from RUFAS.biophysical.field.soil.soil_data import SoilData
+from RUFAS.output_manager import OutputManager
 
 
 # ---- Static method tests
@@ -46,18 +48,20 @@ def test_determine_moisture_factor(dry_fraction: float) -> None:
 
 
 @pytest.mark.parametrize(
-    "mass,dry_fraction",
+    "dry_fraction",
     [
-        (500, 0),
-        (400, -0.5),
-        (600, 1.1),
+        0,
+        -0.5,
+        1.1,
     ],
 )
-def test_error_determine_moisture_factor(mass: float, dry_fraction: float) -> None:
+def test_error_determine_moisture_factor(dry_fraction: float, mocker: MockerFixture) -> None:
     """Tests that correct error is raised when invalid argument is passed."""
+    mock_add_error = mocker.patch.object(OutputManager, "add_error")
     with pytest.raises(ValueError) as e:
         ManureApplication._determine_moisture_factor(dry_fraction)
     assert str(e.value) == f"Dry matter content must be in the range (0.0, 1.0], received: '{dry_fraction}'."
+    mock_add_error.assert_called_once()
 
 
 @pytest.mark.parametrize(
@@ -113,12 +117,14 @@ def test_determine_water_extractable_inorganic_phosphorus_fraction_by_animal(ani
 
 @pytest.mark.parametrize("animal_type", ["CaTTLE", "PORK", "fish"])
 def test_error_determine_water_extractable_inorganic_phosphorus_fraction_by_animal(
-    animal_type: str,
+    animal_type: str, mocker: MockerFixture
 ) -> None:
     """Tests that errors caused by unsupported animal types are handled appropriately."""
+    mock_add_error = mocker.patch.object(OutputManager, "add_error")
     with pytest.raises(ValueError) as e:
         ManureApplication._determine_water_extractable_inorganic_phosphorus_fraction_by_animal(animal_type)
     assert str(e.value) == f'Expected "CATTLE", "SWINE", or "POULTRY", received \'{animal_type}\'.'
+    mock_add_error.assert_called_once()
 
 
 # ---- Helper function tests
@@ -588,11 +594,12 @@ def test_apply_machine_manure(
     weiP_frac: float | None,
     source_animal: str,
     should_fail: bool,
+    mocker: MockerFixture,
 ) -> None:
     """Tests that the machine-applied manure is correctly added into existing manure on the field."""
     data = SoilData(field_size=area)
     incorp = ManureApplication(data)
-
+    mock_add_error = mocker.patch.object(OutputManager, "add_error")
     if should_fail:
         with pytest.raises(ValueError) as e:
             incorp.apply_machine_manure(
@@ -613,6 +620,7 @@ def test_apply_machine_manure(
             str(e.value) == f"Water extractable inorganic phosphorus fraction must be in the range [0.0, 0.95],"
             f" received '{weiP_frac}'."
         )
+        mock_add_error.assert_called_once()
     else:
         incorp._determine_water_extractable_inorganic_phosphorus_fraction_by_animal = MagicMock(return_value=0.25)
         incorp.apply_machine_manure(
@@ -641,3 +649,4 @@ def test_apply_machine_manure(
         assert incorp.data.machine_manure.water_extractable_organic_phosphorus > 0
         assert incorp.data.machine_manure.stable_inorganic_phosphorus > 0
         assert incorp.data.machine_manure.stable_organic_phosphorus > 0
+        mock_add_error.assert_not_called()